requirements.txt
Django==3.2.5
djangorestframework==3.12.4
psycopg2==2.9.6
django-cors-headers==3.7.0
djangorestframework-jwt==1.11.0
requests==2.31.0
urllib3==2.0.3
uuid==1.30
Pillow==8.2.0